pyproject.toml
[tool.poetry]
name = "todoist-cli"
version = "0.1.0"
description = ""
authors = ["makridenko <makridenko.a@yandex.ru>"]
readme = "README.md"
packages = [{include = "todoist_cli"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pre-commit = "^2.20.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
skip-string-normalization = true
